Add Portfolio component and some code refactor
Among the changes made in this commit are the following:

- An improvement in the quality of the written code was
made especially in the css classes looking to use the DRY
methodology.
- The Portfolio component is constituted at the same time
by other sub components that are going to be all the projects
that I have made, looking to show the code in a more readable
and easy to understand way.