Solve AOS library issue
This commit fixed the problem when displaying the website
on mobile and tablets devices, since the animations were
creating a blank space on the right side of the website.
To fix that problem, I added css classes related to the
overflow in the parent components of the elements that
were generating that unexpected behavior.